+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.5.0...HEAD)
 ------------------------------------------------------------------------
+### Changed
+- Updates syntax highlighting
 
 [0.5.0](https://github.com/jshwi/docsig/releases/tag/v0.5.0) - 2022-06-21
 ------------------------------------------------------------------------