+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Updates syntax highlighting
 
+### Fixed
+- Fixes returning of non-subscript types
+
 [0.5.0](https://github.com/jshwi/docsig/releases/tag/v0.5.0) - 2022-06-21
 ------------------------------------------------------------------------
 ### Added