+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/docsig/compare/v0.5.0...HEAD)
+[Unreleased](https://github.com/jshwi/docsig/compare/v0.6.0...HEAD)
+------------------------------------------------------------------------
+
+[0.6.0](https://github.com/jshwi/docsig/releases/tag/v0.6.0) - 2022-06-21
 ------------------------------------------------------------------------
 ### Added
 - Allows `key(word)` for `**kwargs`